Refactor SLS systemd infrastructure
This commit proposes 3 modifications in the systemd infrastructure:

(a) Move the files to systemd/ directory in the source tree, to make
it more clear.

(b) Add a randomization of 3m in the timer, to prevent the service
from always running exactly in the same time, causing maybe a bad
relation with some other services;

(c) Stop using the fork/disown approach from before - now we invoke
directly the relevant script with no usage of the helper script
"submitter-load.sh".

Thanks @endrift / @xexaxo for the feedback/ideas!

Signed-off-by: Guilherme G. Piccoli <gpiccoli@igalia.com>