README: Drop license in the header and rename to txt
As per Vicki suggestion, license in the header of a README
is not necessary. Also, it's a text file in the end, so let's
call it .txt (and not .md).

While at it, stop installing this file in the lib folder,
this belongs to Git repo, doesn't need to install it.

Signed-off-by: Guilherme G. Piccoli <gpiccoli@igalia.com>