helpers/kdump.py: Improve kdump parsing
There are some changes required in this helper in order it collects
more proper stack/note for log submission.

(a) Pstore might collect 2 dmesgs (oops + panic); the data present
in both logs are mostly the same, except the panic one has some more
additional data due to panic_print. Due to the parsing strings we use,
it's required to make use of the panic log, so refactor the logic to
stop parsing when the note/stack is collected, avoid looping in all
dmesgs inside the zip blob.

(b) Due to some changes in upstream kernel, now we have the NMI
backtrace on panic_print showing before the "Kernel Offset" string.
Hence, the dmesg parsing might collect all the CPUs backtrace in the
note section, which is useless. Change here the stop string to be
either the offset or "Sending NMI", the one that comes first.

(c) In the call trace parsing, we need to account "RIP:" two times,
due to the NMI technique used by panic path to collect the oops CPU
call trace. If we just collect up to the first "RIP:" entry, we're
collecting only such NMI trace, which is useless.

(d) Finally, both note and stack contained first/last lines that
are useless and bring no info, so drop them.

Signed-off-by: Guilherme G. Piccoli <gpiccoli@igalia.com>